requirements.txt
Flask
Flask-GraphQL
graphene
pymongo
beautifulsoup4
requests
pillow
Flask-APScheduler
python-dotenv
pytz
gunicorn
Flask-SocketIO
python-socketio
eventlet
